+++ vortex/Cargo.toml
@@ -93,3 +93,7 @@ test = false
 name = "common_encoding_tree_throughput"
 harness = false
 test = false
+
+[[bench]]
+name = "pipeline"
+harness = false
